refactor: Move disablerule path
