+++ internal/addon/rules/internal/visitor/internal/disablerule/command.go
